requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-database-url==2.3.0
Django==5.2.1
django-forms-dynamic==1.0.0
django-simple-history==3.8.0
django-unfold==0.50.0
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
PyPDF2==3.0.1
pypdfium2==4.30.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
whitenoise==6.9.0
